Move setError
